backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.7.0
requests==2.31.0
google-generativeai==0.3.0
